// 04-country-app/src/app/country/interfaces/country.interface.ts
export interface Country {
  cca2: string;
  flag: string;
  flagSvg: string;
  name: string;
  capital: string;
  population: number;
  area: number;
  languages: string[];
  borders: string[];
  region: string;
  subregion: string;
  mapsUbication: string;
  continent: string;
  currencies: string[];
  timezone: string[];
  coatOfArmsSvg: string;
}